+++ firmware/cmake/bootlib.cmake
@@ -52,7 +52,6 @@ function(stm32f4_boot_binary
     # Add shared files.
     list(APPEND SRCS
             "${SHARED_IO_INCLUDE_DIR}/io_canQueue.c"
-            "${SHARED_IO_INCLUDE_DIR}/io_can.c"
             "${SHARED_IO_INCLUDE_DIR}/io_time.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_flash.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_crc.c"
@@ -61,6 +60,7 @@ function(stm32f4_boot_binary
             "${SHARED_HW_INCLUDE_DIR}/hw_crc.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_assert.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_error.c"
+            "${SHARED_HW_INCLUDE_DIR}/hw_can.c"
     )
     list(APPEND INCLUDE_DIRS
             "${SHARED_APP_INCLUDE_DIR}"
@@ -131,7 +131,6 @@ function(stm32h7_boot_binary
     # Add shared files.
     list(APPEND SRCS
             "${SHARED_IO_INCLUDE_DIR}/io_canQueue.c"
-            "${SHARED_IO_INCLUDE_DIR}/io_fdcan.c"
             "${SHARED_IO_INCLUDE_DIR}/io_time.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_flash.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_crc.c"
@@ -140,6 +139,7 @@ function(stm32h7_boot_binary
             "${SHARED_HW_INCLUDE_DIR}/hw_crc.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_assert.c"
             "${SHARED_HW_INCLUDE_DIR}/hw_error.c"
+            "${SHARED_HW_INCLUDE_DIR}/hw_fdcan.c"
     )
     list(APPEND INCLUDE_DIRS
             "${SHARED_APP_INCLUDE_DIR}"
