+++ firmware/quadruna/VC/src/cubemx/VC.ioc
@@ -101,11 +101,11 @@ Dma.USART2_RX.1.SyncPolarity=HAL_DMAMUX_SYNC_NO_EVENT
 Dma.USART2_RX.1.SyncRequestNumber=1
 Dma.USART2_RX.1.SyncSignalID=NONE
 FDCAN1.AutoRetransmission=ENABLE
-FDCAN1.CalculateBaudRateNominal=1000000
-FDCAN1.CalculateTimeBitNominal=1000
-FDCAN1.CalculateTimeQuantumNominal=62.5
+FDCAN1.CalculateBaudRateNominal=500000
+FDCAN1.CalculateTimeBitNominal=2000
+FDCAN1.CalculateTimeQuantumNominal=125.0
 FDCAN1.IPParameters=CalculateTimeQuantumNominal,CalculateTimeBitNominal,CalculateBaudRateNominal,AutoRetransmission,NominalSyncJumpWidth,StdFiltersNbr,RxFifo0ElmtsNbr,TxFifoQueueElmtsNbr,NominalTimeSeg1,NominalTimeSeg2,NominalPrescaler
-FDCAN1.NominalPrescaler=6
+FDCAN1.NominalPrescaler=12
 FDCAN1.NominalSyncJumpWidth=3
 FDCAN1.NominalTimeSeg1=12
 FDCAN1.NominalTimeSeg2=3
@@ -523,7 +523,7 @@ ProjectManager.FreePins=false
 ProjectManager.HalAssertFull=false
 ProjectManager.HeapSize=0x200
 ProjectManager.KeepUserCode=true
-ProjectManager.LastFirmware=false
+ProjectManager.LastFirmware=true
 ProjectManager.LibraryCopy=0
 ProjectManager.MainLocation=Core/Src
 ProjectManager.NoMain=false
