+++ firmware/shared/src/hw/hw_can.c
@@ -1,4 +1,4 @@
-#include "io_can.h"
+#include "hw_can.h"
 #undef NDEBUG // TODO remove this in favour of always_assert (we would write this)
 #include <assert.h>
 #include "io_time.h"
@@ -30,7 +30,7 @@
 #define MASKMODE_16BIT_ID_OPEN INIT_MASKMODE_16BIT_FiRx(0x0, CAN_ID_STD, CAN_RTR_DATA, CAN_ExtID_NULL)
 #define MASKMODE_16BIT_MASK_OPEN INIT_MASKMODE_16BIT_FiRx(0x0, 0x1, 0x1, 0x0)
 
-void io_can_init(const CanHandle *can_handle)
+void hw_can_init(const CanHandle *can_handle)
 {
     // Configure a single filter bank that accepts any message.
     CAN_FilterTypeDef filter;
@@ -57,13 +57,13 @@ void io_can_init(const CanHandle *can_handle)
     assert(HAL_CAN_Start(can_handle->hcan) == HAL_OK);
 }
 
-void io_can_deinit(const CanHandle *can_handle)
+void hw_can_deinit(const CanHandle *can_handle)
 {
     assert(HAL_CAN_Stop(can_handle->hcan) == HAL_OK);
     assert(HAL_CAN_DeInit(can_handle->hcan) == HAL_OK);
 }
 
-bool io_can_transmit(const CanHandle *can_handle, CanMsg *msg)
+bool hw_can_transmit(const CanHandle *can_handle, CanMsg *msg)
 {
     CAN_TxHeaderTypeDef tx_header;
 
@@ -96,7 +96,7 @@ bool io_can_transmit(const CanHandle *can_handle, CanMsg *msg)
     return return_status == HAL_OK;
 }
 
-bool io_can_receive(const CanHandle *can_handle, const uint32_t rx_fifo, CanMsg *msg)
+bool hw_can_receive(const CanHandle *can_handle, const uint32_t rx_fifo, CanMsg *msg)
 {
     CAN_RxHeaderTypeDef header;
     if (HAL_CAN_GetRxMessage(can_handle->hcan, rx_fifo, &header, msg->data) != HAL_OK)
